Fix some tick variable types, rewrite timermanager to use boost more

+++ dcpp/HashManager.h
@@ -235,7 +235,7 @@ private:
 		store.rebuild();
 	}
 
-	virtual void on(TimerManagerListener::Minute, uint32_t) throw() {
+	virtual void on(TimerManagerListener::Minute, uint64_t) throw() {
 		Lock l(cs);
 		store.save();
 	}
